feature(subsetter): metadata catalog and collection/variable search
