+++ src/components/time-series/time-series.component.ts
@@ -17,7 +17,7 @@ import type { CSSResultGroup } from 'lit'
 import type { Plot } from '../plot/plot.types.js'
 import type { MenuNames } from './time-series.types.js'
 import type { Variable } from '../browse-variables/browse-variables.types.js'
-import { GiovanniVariableCatalog } from '../../variable-catalog/giovanni-variable-catalog.js'
+import { GiovanniVariableCatalog } from '../../metadata-catalog/giovanni-variable-catalog.js'
 import { DB_NAME, getDataByKey, IndexedDbStores } from '../../internal/indexeddb.js'
 import type { VariableDbEntry } from './time-series.types.js'
 import type { TerraPlotRelayoutEvent } from '../../events/terra-plot-relayout.js'
