+++ src/metadata-catalog/giovanni-variable-catalog.ts
