+++ src/metadata-catalog/queries.ts
@@ -1,5 +1,64 @@
 import { gql } from '@apollo/client/core'
 
+export const GET_CMR_SEARCH_RESULTS_ALL = gql`
+    query GetCMRSearchResultsAll($keyword: String!) {
+        collections(params: { keyword: $keyword }) {
+            items {
+                conceptId
+                nativeId
+                provider
+                title
+            }
+        }
+        variables(params: { keyword: $keyword }) {
+            items {
+                conceptId
+                name
+                providerId
+                longName
+                collections {
+                    items {
+                        conceptId
+                        nativeId
+                    }
+                }
+            }
+        }
+    }
+`
+
+export const GET_CMR_SEARCH_RESULTS_COLLECTIONS = gql`
+    query GetCMRSearchResultsCollections($keyword: String!) {
+        collections(params: { keyword: $keyword }) {
+            items {
+                conceptId
+                nativeId
+                provider
+                title
+            }
+        }
+    }
+`
+
+export const GET_CMR_SEARCH_RESULTS_VARIABLES = gql`
+    query GetCMRSearchResultsVariables($keyword: String!) {
+        variables(params: { keyword: $keyword }) {
+            items {
+                conceptId
+                name
+                providerId
+                longName
+                collections {
+                    items {
+                        conceptId
+                        nativeId
+                    }
+                }
+            }
+        }
+    }
+`
+
 export const GET_SEARCH_KEYWORDS = gql`
     query {
         aesirKeywords {
